moves checkboxTools to VueJobsTools
